refactor(model): move all OOAPI data structs into ooapi.go
Follows naturally from the previous diff.
